Increase amount of rows to fetch at a time. (#394)
* Increase amount of rows to fetch at a time.
Use constant for fetch count.
Clean up some mis-matched type related compiler warnings

* Fix fetching threshold logic when scrolling.
When scrolling backwards start position was set incorrectly causing fetches to occur every `m_fetch_threshold_items` rows instead of after an entire chunks worth (1000) was scrolled.

+++ src/view/src/widgets/rocprofvis_infinite_scroll_table.h
@@ -50,7 +50,7 @@ protected:
     DataProvider&    m_data_provider;
     SettingsManager& m_settings;
 
-    int m_fetch_chunk_size;
+    uint64_t m_fetch_chunk_size;
 
     bool m_data_changed;
